Refactor CLI structure to modularize command handling and improve code organization
