+++ src/cli/commands/collections.ts
@@ -1,0 +1,8 @@
+import { dspaceClient } from '../services/dspace-client.service'
+
+export const collectionsCommands = {
+    async handleCollectionsList(): Promise<void> {
+        await dspaceClient.ensureAuth()
+        await dspaceClient.showCollections()
+    }
+}
