+++ src/cli/services/dspace-client.service.ts
@@ -1,0 +1,31 @@
+import * as DSpaceClient from '../../index'
+import { configService } from './config.service'
+import { authStore } from '../utils/store'
+
+export const dspaceClient = {
+  init: DSpaceClient.init,
+  login: DSpaceClient.login,
+  showAllItems: DSpaceClient.showAllItems,
+  showItem: DSpaceClient.showItem,
+  updateItem: DSpaceClient.updateItem,
+  showItemBundles: DSpaceClient.showItemBundles,
+  newBitstream: DSpaceClient.newBitstream,
+  deleteBitstreams: DSpaceClient.deleteBitstreams,
+  showCollections: DSpaceClient.showCollections,
+  moveItem: DSpaceClient.moveItem,
+
+  async ensureAuth(): Promise<void> {
+    const config = configService.loadConfig()
+    if (!config.baseURL) {
+      throw new Error('Set baseURL first with config:set baseURL <url>')
+    }
+
+    this.init(config.baseURL as string)
+
+    const credentials = authStore.get('credentials')
+    if (!credentials) {
+      throw new Error('No saved credentials. Run "dspace-cli login" first.')
+    }
+    await this.login(credentials.username, credentials.password)
+  }
+}
