+++ src/cli/types/index.ts
@@ -1,0 +1,5 @@
+export interface Config {
+  baseURL?: string
+  defaultCommunity?: string
+  [key: string]: unknown
+}
